Parselets organization and type parser basics
