+++ src/lexer/SymbolDecypher.php
@@ -51,7 +51,7 @@ class SymbolDecypher
             case '.':
                 return static::tryMatch($context, ['...', '..']);
             case '#':
-                return static::tryMatch($context, ['#{']);
+                return static::tryMatch($context, ['#{', '#(']);
             case '%':
                 return static::tryMatch($context, ['%{']);
             default:
