+++ src/parselets/expr/AccessParselet.php
@@ -19,32 +19,26 @@
  * You should have received a copy of the GNU General Public License
  * along with Quack.  If not, see <http://www.gnu.org/licenses/>.
  */
-namespace QuackCompiler\Parselets;
+namespace QuackCompiler\Parselets\Expr;
 
-use \QuackCompiler\Parser\Precedence;
-use \QuackCompiler\Parser\Grammar;
-use \QuackCompiler\Parser\SyntaxError;
+use \QuackCompiler\Ast\Expr\AccessExpr;
 use \QuackCompiler\Ast\Expr\Expr;
-use \QuackCompiler\Ast\Expr\TernaryExpr;
-use \QuackCompiler\Ast\Expr\NameExpr;
-use \QuackCompiler\Ast\Expr\OperatorExpr;
 use \QuackCompiler\Lexer\Token;
-use \QuackCompiler\Lexer\Tag;
+use \QuackCompiler\Parser\Grammar;
+use \QuackCompiler\Parser\Precedence;
 
-class MemberAccessParselet implements IInfixParselet
+class AccessParselet implements IInfixParselet
 {
     public function parse(Grammar $grammar, Expr $left, Token $token)
     {
-        $right = $grammar->_name();
-        return new OperatorExpr(
-            $left,
-            $token->getTag(),
-            $grammar->parser->resolveScope($right->getPointer())
-        );
+        $index = $grammar->_expr();
+        $grammar->parser->match('}');
+
+        return new AccessExpr($left, $index);
     }
 
     public function getPrecedence()
     {
-        return Precedence::MEMBER_ACCESS;
+        return Precedence::ACCESS;
     }
 }
