+++ src/parselets/expr/IInfixParselet.php
@@ -19,32 +19,14 @@
  * You should have received a copy of the GNU General Public License
  * along with Quack.  If not, see <http://www.gnu.org/licenses/>.
  */
-namespace QuackCompiler\Parselets;
+namespace QuackCompiler\Parselets\Expr;
 
-use \QuackCompiler\Parser\Grammar;
 use \QuackCompiler\Ast\Expr\Expr;
-use \QuackCompiler\Ast\Expr\OperatorExpr;
 use \QuackCompiler\Lexer\Token;
+use \QuackCompiler\Parser\Grammar;
 
-class BinaryOperatorParselet implements IInfixParselet
+interface IInfixParselet
 {
-    public $precedence;
-    public $is_right;
-
-    public function __construct($precedence, $is_right)
-    {
-        $this->precedence = $precedence;
-        $this->is_right = $is_right;
-    }
-
-    public function parse(Grammar $parser, Expr $left, Token $token)
-    {
-        $right = $parser->_expr($this->precedence - ($this->is_right ? 1 : 0));
-        return new OperatorExpr($left, $token->getTag(), $right);
-    }
-
-    public function getPrecedence()
-    {
-        return $this->precedence;
-    }
+    public function parse(Grammar $parser, Expr $left, Token $token);
+    public function getPrecedence();
 }
