+++ src/parselets/expr/IPrefixParselet.php
@@ -19,27 +19,12 @@
  * You should have received a copy of the GNU General Public License
  * along with Quack.  If not, see <http://www.gnu.org/licenses/>.
  */
-namespace QuackCompiler\Parselets;
+namespace QuackCompiler\Parselets\Expr;
 
-use \QuackCompiler\Parser\Precedence;
-use \QuackCompiler\Parser\Grammar;
-use \QuackCompiler\Ast\Expr\Expr;
-use \QuackCompiler\Ast\Expr\TernaryExpr;
 use \QuackCompiler\Lexer\Token;
-use \QuackCompiler\Lexer\Tag;
+use \QuackCompiler\Parser\Grammar;
 
-class TernaryParselet implements IInfixParselet
+interface IPrefixParselet
 {
-    public function parse(Grammar $grammar, Expr $left, Token $token)
-    {
-        $then = $grammar->_expr();
-        $grammar->parser->match(Tag::T_ELSE);
-        $else = $grammar->_expr(Precedence::TERNARY - 1);
-        return new TernaryExpr($left, $then, $else);
-    }
-
-    public function getPrecedence()
-    {
-        return Precedence::TERNARY;
-    }
+    public function parse(Grammar $parser, Token $token);
 }
