+++ src/parser/Parser.php
@@ -25,28 +25,28 @@ use \Exception;
 use \QuackCompiler\Lexer\Tag;
 use \QuackCompiler\Lexer\Token;
 use \QuackCompiler\Lexer\Tokenizer;
-use \QuackCompiler\Parselets\IPrefixParselet;
-use \QuackCompiler\Parselets\IInfixParselet;
-use \QuackCompiler\Parselets\BinaryOperatorParselet;
-use \QuackCompiler\Parselets\LiteralParselet;
-use \QuackCompiler\Parselets\NameParselet;
-use \QuackCompiler\Parselets\PostfixOperatorParselet;
-use \QuackCompiler\Parselets\PrefixOperatorParselet;
-use \QuackCompiler\Parselets\TernaryParselet;
-use \QuackCompiler\Parselets\GroupParselet;
-use \QuackCompiler\Parselets\LambdaParselet;
-use \QuackCompiler\Parselets\ArrayParselet;
-use \QuackCompiler\Parselets\NewParselet;
-use \QuackCompiler\Parselets\MemberAccessParselet;
-use \QuackCompiler\Parselets\WhenParselet;
-use \QuackCompiler\Parselets\CallParselet;
-use \QuackCompiler\Parselets\AccessParselet;
-use \QuackCompiler\Parselets\RangeParselet;
-use \QuackCompiler\Parselets\PartialFuncParselet;
-use \QuackCompiler\Parselets\WhereParselet;
-use \QuackCompiler\Parselets\MapParselet;
-use \QuackCompiler\Parselets\ObjectParselet;
-use \QuackCompiler\Parselets\BlockParselet;
+use \QuackCompiler\Parselets\Expr\IPrefixParselet;
+use \QuackCompiler\Parselets\Expr\IInfixParselet;
+use \QuackCompiler\Parselets\Expr\BinaryOperatorParselet;
+use \QuackCompiler\Parselets\Expr\LiteralParselet;
+use \QuackCompiler\Parselets\Expr\NameParselet;
+use \QuackCompiler\Parselets\Expr\PostfixOperatorParselet;
+use \QuackCompiler\Parselets\Expr\PrefixOperatorParselet;
+use \QuackCompiler\Parselets\Expr\TernaryParselet;
+use \QuackCompiler\Parselets\Expr\GroupParselet;
+use \QuackCompiler\Parselets\Expr\LambdaParselet;
+use \QuackCompiler\Parselets\Expr\ArrayParselet;
+use \QuackCompiler\Parselets\Expr\NewParselet;
+use \QuackCompiler\Parselets\Expr\MemberAccessParselet;
+use \QuackCompiler\Parselets\Expr\WhenParselet;
+use \QuackCompiler\Parselets\Expr\CallParselet;
+use \QuackCompiler\Parselets\Expr\AccessParselet;
+use \QuackCompiler\Parselets\Expr\RangeParselet;
+use \QuackCompiler\Parselets\Expr\PartialFuncParselet;
+use \QuackCompiler\Parselets\Expr\WhereParselet;
+use \QuackCompiler\Parselets\Expr\MapParselet;
+use \QuackCompiler\Parselets\Expr\ObjectParselet;
+use \QuackCompiler\Parselets\Expr\BlockParselet;
 
 abstract class Parser
 {
