+++ src/toolkit/QuackToolkit.php
@@ -60,28 +60,28 @@ import(PARSER, 'SyntaxError');
 import(PARSER, 'TokenReader');
 import(PARSER, 'Precedence');
 
-import(PARSELETS, 'IInfixParselet');
-import(PARSELETS, 'IPrefixParselet');
-import(PARSELETS, 'PrefixOperatorParselet');
-import(PARSELETS, 'BinaryOperatorParselet');
-import(PARSELETS, 'PostfixOperatorParselet');
-import(PARSELETS, 'TernaryParselet');
-import(PARSELETS, 'GroupParselet');
-import(PARSELETS, 'LambdaParselet');
-import(PARSELETS, 'ArrayParselet');
-import(PARSELETS, 'NameParselet');
-import(PARSELETS, 'NewParselet');
-import(PARSELETS, 'MemberAccessParselet');
-import(PARSELETS, 'WhenParselet');
-import(PARSELETS, 'CallParselet');
-import(PARSELETS, 'AccessParselet');
-import(PARSELETS, 'RangeParselet');
-import(PARSELETS, 'LiteralParselet');
-import(PARSELETS, 'PartialFuncParselet');
-import(PARSELETS, 'WhereParselet');
-import(PARSELETS, 'MapParselet');
-import(PARSELETS, 'ObjectParselet');
-import(PARSELETS, 'BlockParselet');
+import(PARSELETS, 'expr/IInfixParselet');
+import(PARSELETS, 'expr/IPrefixParselet');
+import(PARSELETS, 'expr/PrefixOperatorParselet');
+import(PARSELETS, 'expr/BinaryOperatorParselet');
+import(PARSELETS, 'expr/PostfixOperatorParselet');
+import(PARSELETS, 'expr/TernaryParselet');
+import(PARSELETS, 'expr/GroupParselet');
+import(PARSELETS, 'expr/LambdaParselet');
+import(PARSELETS, 'expr/ArrayParselet');
+import(PARSELETS, 'expr/NameParselet');
+import(PARSELETS, 'expr/NewParselet');
+import(PARSELETS, 'expr/MemberAccessParselet');
+import(PARSELETS, 'expr/WhenParselet');
+import(PARSELETS, 'expr/CallParselet');
+import(PARSELETS, 'expr/AccessParselet');
+import(PARSELETS, 'expr/RangeParselet');
+import(PARSELETS, 'expr/LiteralParselet');
+import(PARSELETS, 'expr/PartialFuncParselet');
+import(PARSELETS, 'expr/WhereParselet');
+import(PARSELETS, 'expr/MapParselet');
+import(PARSELETS, 'expr/ObjectParselet');
+import(PARSELETS, 'expr/BlockParselet');
 
 /* Ast */
